fluxgate::config

Function build_middleware

fn build_middleware<'b>(
    analytics_channel: Option<Arc<Mutex<ChildStdin>>>,
    server_name: Arc<str>,
) -> impl for<'a> Service<&'a [u8], Error = MiddlewareError, Response = ()> + Clone + Send + use<'b>

Builds a middleware for the server using the provided analytics channel and server name.

Arguments

analytics_channel: An optional channel for analytics.

server_name: The name of the server.

Returns

A middleware service that can be used for processing requests.