fluxgate::controller

Function main_loop

async fn main_loop<T>(
    transport: T,
    server_addr: SocketAddr,
    early_data: Vec<u8>,
    middleware: impl for<'a> Service<&'a [u8], Error = MiddlewareError>,
) -> Result<()>
where
    T: AsyncRead + AsyncWrite + Unpin,

Main loop for handling the proxying of data between the client and the server.

Arguments

transport - The transport stream.

server_addr - The server address.

early_data - The early data to send to the server.

Returns

An io::Result indicating the success or failure of the operation.